feat(chat): extract ChatThread component from AdminChatSheet
Separate thread view (messages, WebSocket, input) into reusable ChatThread.tsx.
AdminChatSheet now manages view state only, delegating thread rendering to ChatThread.

Refs: fresh-catch-mol-p10j
