+++ src/app/pages/admin/chat/index.ts
@@ -1,2 +1,3 @@
 export { AdminChatBubble } from './AdminChatBubble';
 export { AdminChatSheet } from './AdminChatSheet';
+export { ChatThread } from './ChatThread';
